# command_center/backend/openapi.json
{
  "components": {
    "schemas": {
...
      "BetaResponse": {
        "properties": {
          "alpha": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Alpha"
          },
          "benchmark": {
            "title": "Benchmark",
            "type": "string"
          },
          "beta": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Beta"
          },
          "meta": {
            "$ref": "#/components/schemas/SeriesMeta"
          },
          "n": {
            "title": "N",
            "type": "integer"
          },
          "r_squared": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "R Squared"
          },
          "rolling": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/Point"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rolling"
          },
          "ticker": {
            "title": "Ticker",
            "type": "string"
          }
        },
        "required": [
          "ticker",
          "benchmark",
          "beta",
          "alpha",
          "r_squared",
          "n",
          "meta"
        ],
        "title": "BetaResponse",
        "type": "object"
      },
...
      "CorrelationResponse": {
        "properties": {
          "clustered": {
            "title": "Clustered",
            "type": "boolean"
          },
          "matrix": {
            "items": {
              "items": {
                "anyOf": [
                  {
                    "type": "number"
                  },
                  {
                    "type": "null"
                  }
                ]
              },
              "type": "array"
            },
            "title": "Matrix",
            "type": "array"
          },
          "meta": {
            "$ref": "#/components/schemas/SeriesMeta"
          },
          "method": {
            "title": "Method",
            "type": "string"
          },
          "tickers": {
            "items": {
              "type": "string"
            },
            "title": "Tickers",
            "type": "array"
          }
        },
        "required": [
          "tickers",
          "matrix",
          "method",
          "clustered",
          "meta"
        ],
        "title": "CorrelationResponse",
        "type": "object"
      },
...
      "DistributionResponse": {
        "properties": {
          "drawdown": {
            "items": {
              "$ref": "#/components/schemas/Point"
            },
            "title": "Drawdown",
            "type": "array"
          },
          "histogram": {
            "$ref": "#/components/schemas/HistogramOut"
          },
          "kind": {
            "title": "Kind",
            "type": "string"
          },
          "meta": {
            "$ref": "#/components/schemas/SeriesMeta"
          },
          "moments": {
            "$ref": "#/components/schemas/MomentsOut"
          },
          "normal": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/NormalOut"
              },
              {
                "type": "null"
              }
            ]
          },
          "qq": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/QQOut"
              },
              {
                "type": "null"
              }
            ]
          },
          "ticker": {
            "title": "Ticker",
            "type": "string"
          }
        },
        "required": [
          "ticker",
          "kind",
          "histogram",
          "moments",
          "drawdown",
          "meta"
        ],
        "title": "DistributionResponse",
        "type": "object"
      },
      "Field": {
        "enum": [
          "open",
          "high",
          "low",
          "close",
          "adj_close",
          "volume"
        ],
        "title": "Field",
        "type": "string"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "PerformanceResponse": {
        "properties": {
          "base": {
            "title": "Base",
            "type": "number"
          },
          "meta": {
            "$ref": "#/components/schemas/SeriesMeta"
          },
          "series": {
            "items": {
              "$ref": "#/components/schemas/NamedSeries"
            },
            "title": "Series",
            "type": "array"
          },
          "tickers": {
            "items": {
              "type": "string"
            },
            "title": "Tickers",
            "type": "array"
          }
        },
        "required": [
          "tickers",
          "base",
          "series",
          "meta"
        ],
        "title": "PerformanceResponse",
        "type": "object"
      },
      "Point": {
        "properties": {
          "t": {
            "title": "T",
            "type": "integer"
          },
          "v": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "V"
          }
        },
        "required": [
          "t",
          "v"
        ],
        "title": "Point",
        "type": "object"
      },
...
      "ReturnKind": {
        "enum": [
          "simple",
          "log"
        ],
        "title": "ReturnKind",
        "type": "string"
      },
      "ReturnsResponse": {
        "properties": {
          "kind": {
            "title": "Kind",
            "type": "string"
          },
          "meta": {
            "$ref": "#/components/schemas/SeriesMeta"
          },
          "series": {
            "items": {
              "$ref": "#/components/schemas/Point"
            },
            "title": "Series",
            "type": "array"
          },
          "ticker": {
            "title": "Ticker",
            "type": "string"
          }
        },
        "required": [
          "ticker",
          "kind",
          "series",
          "meta"
        ],
        "title": "ReturnsResponse",
        "type": "object"
      },
      "RollingResponse": {
        "properties": {
          "annualized": {
            "title": "Annualized",
            "type": "boolean"
          },
          "meta": {
            "$ref": "#/components/schemas/SeriesMeta"
          },
          "series": {
            "items": {
              "$ref": "#/components/schemas/Point"
            },
            "title": "Series",
            "type": "array"
          },
          "stat": {
            "title": "Stat",
            "type": "string"
          },
          "ticker": {
            "title": "Ticker",
            "type": "string"
          },
          "window": {
            "title": "Window",
            "type": "integer"
          }
        },
        "required": [
          "ticker",
          "stat",
          "window",
          "annualized",
          "series",
          "meta"
        ],
        "title": "RollingResponse",
        "type": "object"
      },
      "RollingStat": {
        "enum": [
          "vol",
          "mean",
          "sharpe"
        ],
        "title": "RollingStat",
        "type": "string"
      },
...
    }
  },
  "info": {
    "description": "Bloomberg-style research terminal API over quant_core.",
    "title": "command_center",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/v1/beta/{ticker}": {
      "get": {
        "operationId": "getBeta",
        "parameters": [
          {
            "in": "path",
            "name": "ticker",
            "required": true,
            "schema": {
              "pattern": "^[A-Z0-9.\\-]{1,15}$",
              "title": "Ticker",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "benchmark",
            "required": false,
            "schema": {
              "default": "SPY",
              "title": "Benchmark",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "start",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Start"
            }
          },
          {
            "in": "query",
            "name": "end",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "End"
            }
          },
          {
            "in": "query",
            "name": "window",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "minimum": 2,
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Window"
            }
          },
          {
            "in": "query",
            "name": "kind",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/ReturnKind",
              "default": "log"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BetaResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Beta"
      }
    },
...
    "/api/v1/correlation": {
      "get": {
        "operationId": "getCorrelation",
        "parameters": [
          {
            "description": "comma-separated tickers",
            "in": "query",
            "name": "tickers",
            "required": true,
            "schema": {
              "description": "comma-separated tickers",
              "title": "Tickers",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "start",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Start"
            }
          },
          {
            "in": "query",
            "name": "end",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "End"
            }
          },
          {
            "in": "query",
            "name": "kind",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/ReturnKind",
              "default": "log"
            }
          },
          {
            "in": "query",
            "name": "field",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/Field",
              "default": "adj_close"
            }
          },
          {
            "in": "query",
            "name": "method",
            "required": false,
            "schema": {
              "default": "pearson",
              "title": "Method",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "cluster",
            "required": false,
            "schema": {
              "default": true,
              "title": "Cluster",
              "type": "boolean"
            }
          },
          {
            "in": "query",
            "name": "interval",
            "required": false,
            "schema": {
              "default": "1d",
              "title": "Interval",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CorrelationResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Correlation"
      }
    },
    "/api/v1/distribution/{ticker}": {
      "get": {
        "operationId": "getDistribution",
        "parameters": [
          {
            "in": "path",
            "name": "ticker",
            "required": true,
            "schema": {
              "pattern": "^[A-Z0-9.\\-]{1,15}$",
              "title": "Ticker",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "start",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Start"
            }
          },
          {
            "in": "query",
            "name": "end",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "End"
            }
          },
          {
            "in": "query",
            "name": "kind",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/ReturnKind",
              "default": "log"
            }
          },
          {
            "in": "query",
            "name": "field",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/Field",
              "default": "adj_close"
            }
          },
          {
            "in": "query",
            "name": "bins",
            "required": false,
            "schema": {
              "default": "fd",
              "title": "Bins",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "qq",
            "required": false,
            "schema": {
              "default": true,
              "title": "Qq",
              "type": "boolean"
            }
          },
          {
            "in": "query",
            "name": "normality",
            "required": false,
            "schema": {
              "default": true,
              "title": "Normality",
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DistributionResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Distribution"
      }
    },
...
    "/api/v1/performance": {
      "get": {
        "operationId": "getPerformance",
        "parameters": [
          {
            "description": "comma-separated tickers",
            "in": "query",
            "name": "tickers",
            "required": true,
            "schema": {
              "description": "comma-separated tickers",
              "title": "Tickers",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "start",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Start"
            }
          },
          {
            "in": "query",
            "name": "end",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "End"
            }
          },
          {
            "in": "query",
            "name": "field",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/Field",
              "default": "adj_close"
            }
          },
          {
            "in": "query",
            "name": "base",
            "required": false,
            "schema": {
              "default": 100.0,
              "title": "Base",
              "type": "number"
            }
          },
          {
            "in": "query",
            "name": "interval",
            "required": false,
            "schema": {
              "default": "1d",
              "title": "Interval",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PerformanceResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Performance"
      }
    },
...
    "/api/v1/returns/{ticker}": {
      "get": {
        "operationId": "getReturns",
        "parameters": [
          {
            "in": "path",
            "name": "ticker",
            "required": true,
            "schema": {
              "pattern": "^[A-Z0-9.\\-]{1,15}$",
              "title": "Ticker",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "start",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Start"
            }
          },
          {
            "in": "query",
            "name": "end",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "End"
            }
          },
          {
            "in": "query",
            "name": "kind",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/ReturnKind",
              "default": "log"
            }
          },
          {
            "in": "query",
            "name": "field",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/Field",
              "default": "adj_close"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReturnsResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Returns"
      }
    },
    "/api/v1/rolling/{ticker}": {
      "get": {
        "operationId": "getRolling",
        "parameters": [
          {
            "in": "path",
            "name": "ticker",
            "required": true,
            "schema": {
              "pattern": "^[A-Z0-9.\\-]{1,15}$",
              "title": "Ticker",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "start",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Start"
            }
          },
          {
            "in": "query",
            "name": "end",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "End"
            }
          },
          {
            "in": "query",
            "name": "window",
            "required": false,
            "schema": {
              "default": 21,
              "minimum": 2,
              "title": "Window",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "stat",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/RollingStat",
              "default": "vol"
            }
          },
          {
            "in": "query",
            "name": "annualize",
            "required": false,
            "schema": {
              "default": true,
              "title": "Annualize",
              "type": "boolean"
            }
          },
          {
            "in": "query",
            "name": "rf",
            "required": false,
            "schema": {
              "default": 0.0,
              "title": "Rf",
              "type": "number"
            }
          },
          {
            "in": "query",
            "name": "kind",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/ReturnKind",
              "default": "log"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RollingResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Rolling"
      }
    },
...
  }
}